tinkerbell: defer client creation to first resource operation
As Terraform was not originally designed to Terraform code feed and
configure another provider, if one uses interpolation in provider
configuration, which is a legitimate use (e.g. when you create server
and deploy Tinkerbell using the same Terraform stack which later on adds
hardware entries), client creation will fail, as it is not guaranteed
that the fields will be populated on time.

This commit defers Tinkerbell client creation to the point where first
resource is actually created, which can be ensured with Terraform
depends_on option on the resources level.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>